+++ lib/input/generate.go
@@ -31,7 +31,7 @@ func init() {
 			return NewAsyncReader(TypeGenerate, true, b, log, stats)
 		}),
 		Version: "3.40.0",
-		Status:  docs.StatusBeta,
+		Status:  docs.StatusStable,
 		Summary: `
 Generates messages at a given interval using a [Bloblang](/docs/guides/bloblang/about)
 mapping executed without a context. This allows you to generate messages for